+++ src/App.tsx
@@ -1,29 +1,12 @@
-import { useState } from 'react';
-import { ThemeProvider } from 'styled-components';
-import { darkTheme, GlobalStyles, lightTheme } from './styles/theme';
+import ThemeButton from './components/ThemeButton';
+import { GlobalStyles } from './styles/theme';
 
 function App() {
-  const [theme, setTheme] = useState('light');
-  const isDarkTheme = theme === 'dark';
-  const toggleTheme = () => setTheme(isDarkTheme ? 'light' : 'dark');
   return (
-    <ThemeProvider theme={isDarkTheme ? darkTheme : lightTheme}>
-      <div className='App'>
-        <GlobalStyles />
-        <button onClick={toggleTheme}>
-          {isDarkTheme ? (
-            <span aria-label='Light mode' role='img'>
-              🌞
-            </span>
-          ) : (
-            <span aria-label='Dark mode' role='img'>
-              🌜
-            </span>
-          )}
-        </button>
-        <p>COUCOU</p>
-      </div>
-    </ThemeProvider>
+    <div className='App'>
+      <ThemeButton />
+      <p>COUCOU</p>
+    </div>
   );
 }
