+++ src/styles/theme.ts
@@ -3,16 +3,19 @@ import { createGlobalStyle } from 'styled-components';
 export const lightTheme = {
   body: '#f1f1f1',
   text: '#121620',
+  border: '#121620',
 };
 export const darkTheme = {
   body: '#121620',
   text: '#f1f1f1',
+  border: '#f1f1f1',
 };
 
 export const GlobalStyles = createGlobalStyle`
   body {
     background: ${({ theme }) => theme.body};
     color: ${({ theme }) => theme.text};
+    border: ${({ theme }) => theme.border};;
     transition: background 0.2s ease-in, color 0.2s ease-in;
   }
 `;
